requirements.txt
attrs==25.3.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
docutils==0.21.2
Faker==37.0.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
MarkupSafe==3.0.2
packaging==24.2
pages==0.3
piglet-templates==1.3.2
pluggy==1.5.0
pyee==12.1.1
pyparsing==3.2.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
